[FEAT] add Passport config file
[why]
local Strategy will be added
for extensibility

[what]
create Passport dir
add index.js as passportConfig (config strategies)
add githubStrategy.js file
add token.js (for allocating token, verifing token)

[etc]
etc...
