+++ Backend/Passport/index.js
@@ -1,0 +1,5 @@
+const GithubStrategy = require('./github');
+
+module.exports = () => {
+  GithubStrategy();
+};
